Initial checkin for Sikuli integration.
