--- chorus-sikulix/src/test/junit/org/chorusbdd/chorus/sikulix/sendPanel.py
@@ -1,16 +1,0 @@
-class sendPanel():
-
-    def clickReject(self):
-        print("clickReject")
-
-    def clickSendBid(self):
-        print("clickSendBid")
-
-    def setBid(self, amount):
-        print("setBid" + amount)
-
-    def setBidAndDescription(self, amount, description):
-        print("setBid" + amount + " : " + description)
-
-    def getBid(self):
-        return 21;
