+++ chorus-sikulix/src/test/junit/org/chorusbdd/chorus/sikulix/stepRegexTest.py
@@ -1,0 +1,16 @@
+class stepRegexTest():
+
+    def clickReject(self):
+        print("clickReject")
+
+    def clickSendBid(self):
+        print("clickSendBid")
+
+    def setBid(self, amount):
+        print("setBid" + amount)
+
+    def setBidAndDescription(self, amount, description):
+        print("setBid" + amount + " : " + description)
+
+    def getBid(self):
+        return 21;
