+++ chorus-sikulix/src/test/sikulis/rfq/valueEntry.sikuli/valueEntry.py
@@ -1,0 +1,10 @@
+from sikuli.Sikuli import *
+
+class valueEntry():
+
+    def clickPlus(self):
+        find("spinner.png").click("spinnerPlus.png")
+
+  
+
+#valueEntry().clickPlus()
